feat(state): record_id 升级 64-bit 防 overflow + started_at 改可读格式
两组改动（构建 0 警告、ctest 23/23 绿）：

1) record_id 由 int 改 long long，防长期运行溢出
- types.h/state.c/state.h/reinject.c/reinject.h/output.c/output.h/dispatch.c：
  record_id 及相关 API (state_add/find_by_params/find_by_id/mark_inactive、
  result_ok、reinject_find_overlap、clean_one_record 等) 改 long long
- state_add 加 DCAT_MAX_ID 上限保护，溢出返回 -1
- cJSON 序列化用 (double) 显式转换、load 用 valuedouble 回填 next_id/record_id
- 格式串 %d → %lld；test.h ASSERT_INT_EQ 用 long long

2) started_at 由 unix 时间戳改 "YYYY-MM-DD HH:MM:SS" 本地时间字符串
- types.h: long started_at → char started_at[20]
- state.c: fmt_started_at 助手(localtime_r+strftime)；state_add 格式化；
  state_save 写字符串；state_load 读字符串并兼容旧数字格式(自动转换)
- dispatch.c: dcat query(无 uid) 输出 started_at 也用字符串
- test_state.c: 扩展往返测试(格式校验+保留) + 新增旧数字兼容测试

验证：~/.demoncat/state.json 旧数字记录加载后自动转为可读字符串。

+++ src/core/reinject.h
@@ -16,6 +16,6 @@ int cores_intersect(const unsigned char a[16], const unsigned char b[16]);
  * out_ids[] 仅写入前 max_ids 个(按 max_ids 截断)；返回重叠总数(可能 > max_ids)。
  * 调用方循环须以 min(返回值, max_ids) 为上限；典型调用 max_ids=DCAT_MAX_RECORDS。 */
 int reinject_find_overlap(const fault_def_t *f, const params_t *new_params,
-                          int *out_ids, int max_ids);
+                          long long *out_ids, int max_ids);
 
 #endif
